requirements.txt
Flask==2.1.1
requests==2.26.0
python-dotenv==0.19.1
selenium==4.4.3
beautifulsoup4==4.10.0
validators==0.20.0